Inside variance computation

The following three functions are used to calculate the variance of th difference of two RMSTs. `survw_integratef` is used for the integrations; `inside_var` calculates the expression inside the integral; finally, `var_f` computes the variance.

inside_var(t, ascale_r, ascale_nr, tau, bshape, ascale_cens, p)

Arguments

t
time at which the survival distribution is evaluated

ascale_r
scale parameter for the Weibull distribution for responders

ascale_nr
scale parameter for the Weibull distribution for non-responders

tau
follow-up

bshape
shape parameter for the Weibull distribution

ascale_cens
distributional parameter for the exponential distribution for the censoring

p
event rate for the response

Value

Variance computation